--- types/insight.d.ts
@@ -1,10 +1,0 @@
-export interface UnspentTransaction {
-    address: string,
-    txid: string;
-    vout: number;
-    scriptPubKey: string;
-    amount: number;
-    satoshis: number;
-    height: number;
-    confirmations: number;
-}
